requirements.txt
alembic
psycopg2-binary
Flask
Flask-SQLAlchemy
flask-swagger-ui
python-dotenv
psycopg2
validate-docbr
